pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.3",
    "dnspython>=2.7.0",
    "jinja2>=3.1.6",
    "python-nmap>=0.7.1",
    "python-whois>=0.9.5",
    "requests>=2.32.3",
]
